feat: create interface folder includes all interfaces&validation folder includes schemaes relates#80
