+++ client/src/pages/Client/index.tsx
@@ -3,18 +3,14 @@ import { useState } from 'react';
 import { Button } from '@mui/material';
 import { JobForm } from '../../components';
 import ProposalJob from '../../components/proposalJob/index';
-
-interface Proposal {
-  username: string
-  description: string,
-  attachments: string,
-}
+import { Proposal } from '../../interfaces';
 
 const proposal: Proposal = {
   username: 'Ahmed Safi',
   description: 'Lorem ipsum dolor, sit amet consectetur adipisicing elit. Adipisci quis soluta ipsa minima.',
   attachments: 'https://www.freelancer.com/u/eTranslators?w=f&ngsw-bypass=',
 };
+
 function Client() {
   const { id } = useParams();
   const [showModel, setShowModel] = useState(false);
